pyproject.toml
[project]
name = "unknown-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.7.0",
    "chainlit>=2.5.5",
    "crawl4ai>=0.6.3",
    "google-genai>=1.19.0",
    "huggingface-hub>=0.32.4",
    "mcp>=1.9.3",
    "playwright>=1.52.0",
    "pymysql>=1.1.1",
    "python-dotenv>=1.1.0",
    "semantic-kernel[google,hugging-face,mcp]>=1.28.1",
    "sentence-transformers>=4.1.0",
    "streamlit>=1.45.1",
    "transformers>=4.52.4",
]
